macOS: fix Kerberos tests where possible
This commit fixes Kerberos tests on OS X where possible, or excludes
them from the build.  The issues that can't be worked around only appear
to affect tests of invalid authentication, so hopefully they will be
limited to the Kerberos-specific tests and will not require further
workarounds for more general purpose tests when security is enabled.

Change-Id: I569a5c94b4aeba34b03690906fa1e356b55d9688
Reviewed-on: http://gerrit.cloudera.org:8080/5005
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins