python: allow passing single strings instead of lists in more places
Several of our APIs take lists of column names, but it's easy to
accidentally just pass a string. If we naively iterate over the string,
we end up interpreting "foo" as ["f", "o", "o"] which is obviously
incorrect.

This patch takes the approach of detecting the string argument and
converting to a singleton list.

An alternative would have been to raise TypeError, but it seems like we
do this conversion in other places, so I decided to be consistent.

Change-Id: I1a81dea5356b66b8860d22f9ee2935072fd4cd6c
Reviewed-on: http://gerrit.cloudera.org:8080/5019
Tested-by: Kudu Jenkins
Reviewed-by: Jordan Birdsell <jtbirdsell@apache.org>