C++ client: hide deflate symbols
These snuck in via commit 8a4bb33. Oddly, client_symbol-test.sh was only
observed to be failing on el6 (not on Ubuntu 14.04 or 16.04). The running
theory is it's because the devtoolset linker is old enough that it can't
figure out that these symbols aren't callable by any client code path.

Change-Id: Ibb5b6cf186a9a18583c4695b134c0729b81a4c01
Reviewed-on: http://gerrit.cloudera.org:8080/5023
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins