sasl: don't use the 'client_mech_list' SASL option
This option allows the client to specify which mechanisms it's willing
to negotiate. It turns out the option was added in Cyrus SASL 2.1.25,
which isn't available on el6.

While that option is useful, it's not doing anything magic: it just
filters the list of mechanisms that we pass from the server. This patch
just implements this filtering on our side.

This fixes the TestSaslRpc.TestNoMatchingMechanisms test case on el6.

Change-Id: I570a91e176803fdd0e4324e6f4443b8297ad0395
Reviewed-on: http://gerrit.cloudera.org:8080/5039
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>