sasl: disable error messages assertions on el6
MIT krb5 on RHEL 6 has a bug which means that many of the error messages
from GSSAPI don't get exposed in a reasonable way. So, assertions in
sasl_rpc-test have been failing.

I couldn't find any way to workaround the actual bug and get better
error messages, so this patch just disables those assertions on krb5
1.10 or earlier.

I used a hack to determine the version of Kerberos since spending the
effort to learn enough cmake to do so didn't seem worth the effort for a
test-local purpose.

Change-Id: I1e780e2060dceb322bbbbb772d06b68271ddf675
Reviewed-on: http://gerrit.cloudera.org:8080/5046
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>