[python] - Enable Python 3 Jenkins Build/Test
The current Jenkins flow only builds and executes the unit tests
for Python 2. Since Python 3 is supported we should also setup Jenkins
to build and test Python 3. This patch enables that as well as removes
the multiprocessing test dependency as that has been included as a standard
library since Python 2.6.

Change-Id: I29a15d7b877d1d5d5620bebbf945de68bfb75177
Reviewed-on: http://gerrit.cloudera.org:8080/5059
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins