KUDU-1745. Avoid crashing during failed master lookup
This addresses a SEGV seen when handling a simultaneous failure talking
to both the master and the tablet servers. In this case, the failed
tablet server object is NULL.

This is only a partial fix/workaround - it avoids the crash but it seems
like the retry backoff is not working as expected. Nonetheless, it is at
least an improvement.

Change-Id: I0a06c31c5ad4e26928a7fcabf24f9fd7a02a7e38
Reviewed-on: http://gerrit.cloudera.org:8080/5089
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins