sasl: add a leak suppression for krb5 1.10
This suppresses a known memory leak in krb5 < 1.12 in which a single
error message structure would be leaked in the case that GSSAPI had any
issues acquiring credentials. I bisected this issue against the krb5 git
repository and found the fix is in commit
379d39c17b8930718e98185a5b32a0f7f3e3b4b6 ("Fix a small memory leak in
util_errmap")

I chose to use the suppression file instead of a scoped leak suppression
object since it's still useful to catch any other leaks we might cause
in our use of SASL APIs.

Change-Id: Ied53bf43e98fd622026c00226eb38550bbbbdb38
Reviewed-on: http://gerrit.cloudera.org:8080/5101
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins