KUDU-1722 - Range Partition Failure on UNIXTIME_MICROS
Errors are returned when attempting to add range partition bounds
for a unixtime_micros value when specifying bound types that are
not the default. This is happening because the logic to convert the
bounds to defaults (inclusive for Lower and exclusive for Upper)
converts the UNIXTIME_MICROS value to an INT64. This patch addresses
this issue and includes a test.

Change-Id: I47d23d184a6b73c5c0cdd04519821c4e7eb53ecb
Reviewed-on: http://gerrit.cloudera.org:8080/5119
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>