KUDU-1634 (part 3). Changed tmp files' infix from ".tmp" to ".kudutmp"
As these ".tmp" strings were scattered over the code, I've tried to remove
this duplication as well by defining the single kTmpInfix in path_util.h.
In a couple of places it led to string concatenations, but I suppose it's
not that bad for overall performance?
The necessity of moving from "tmp" to "kudutmp" is backed by the idea that
we don't want to remove any user's data accidentally. For example, we may
have a symlink to some external directory, which doesn't belong only to
Kudu. Hence, we don't want to delete any ordinary tmp files from it.
(see https://gerrit.cloudera.org/#/c/5007/).

Change-Id: I52a794cc16d008570039d276ab32e6a26829bc53
Reviewed-on: http://gerrit.cloudera.org:8080/5123
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>