Fix list rendering of the last weekly update post
The sublist after "Noteworthy features/improvements" is not rendering
properly. Apparently it was missing a new line to be indentified as a
list but adding it made things look weird too (a new line before the
list but none after). This makes the list a "definition list" which
boldens the title and avoids the new line weirdness at the cost of
extra indenting.

Change-Id: I78abc84ee430c0c73536ebc5691c39c47f836498
Reviewed-on: http://gerrit.cloudera.org:8080/5131
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: David Ribeiro Alves <dralves@apache.org>