Fix flakiness of MiniKDC
We use 'lsof' to find the ephemeral port assigned to the MiniKDC, but we
weren't passing the '-P' flag which inhibits the conversion of numeric
port numbers to service names. As such, occasionally the MiniKDC would
be assigned some port which matches a named service in /etc/services and
result in lsof output like:
  p934
  f7
  n*:mediabox
(mediabox is registered as port 46999)

This would cause the port number parsing to fail, making tests ~1%
flaky.

Passing the '-P' flag inhibits the conversion of port numbers to port
names.

Change-Id: I88c0459d4b8ce0371f7cdf5cc38879ecc8cbf5ff
Reviewed-on: http://gerrit.cloudera.org:8080/5156
Reviewed-by: Jordan Birdsell <jtbirdsell@apache.org>
Tested-by: Kudu Jenkins