KUDU-1189 integration test for reusing snap timestamp
Added a test for reusing snapshot timestamp when not set while running
scans in READ_AT_SNAPSHOT mode. This is a test for the functionality
introduced in the context of KUDU-1189.

The test is disabled as it currently fails. A follow up patch will fix
the bug and enable the test.

Change-Id: I7282976580cc15ef330871a838bbf7e46230ceb6
Reviewed-on: http://gerrit.cloudera.org:8080/5163
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins