KUDU-1448 (take 2). Enable AVX2 bitshuffle at runtime
This uses some trickery when building bitshuffle so that we build it
twice: once with normal flags and a second time with -mavx2. Enabling
AVX2 turns on some faster code paths which have been previously
benchmarked to yield a ~30% end-to-end improvement in some Impala scans.

In order to avoid multiply-defined symbols, the build uses 'objcopy' to
rename the symbols to add a '_avx2' suffix before linking. Then a new
wrapper function calls the correct variant of the function based on the
current CPU.

A previous version of this patch tried to use the 'ifunc' attribute, but
ran into some issues with ld.so on el6, likely caused by this bug:
https://patchwork.ozlabs.org/patch/506223/

I'd previously attempted to do this upstream in the bitshuffle library
but it was somewhat difficult as the library makes heavy use of macros,
etc. In that prior attempt, I gave up after many hours of work, whereas
this was comparatively quite simple (only an hour of work!)

This also fixes a bug in the cpuid detection which we inherited
from Chromium. I filed the bug upstream here:
https://bugs.chromium.org/p/chromium/issues/detail?id=667457

I tested on my laptop which has AVX2 and verified that the AVX2 variant
functions were getting called.

Change-Id: Ic115040f5bcdf3aa12a285ad23704be1b0bcb7c6
Previously-Reviewed-on: http://gerrit.cloudera.org:8080/5166
Reviewed-on: http://gerrit.cloudera.org:8080/5247
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>