[tools] Tombstone the tablet with "local_replica delete"
This change makes the default action of 'local_replica delete'
tool to tombstone the tablet. Release 1.1 shipped the tool
with "--clean_unsafe" flag which let you delete the replica
from the disks. Deleting removes consensus metadata for the given
tablet and hence voilates Raft vote durability requirements.
Tombstoning on the other hand will preserve Raft votes for the
specfied tablet on the local node, hence is a safer choice
to be the default action for this tool.

This adds the support for tombstone action for this tool
which was unsupported in previous release.

Change-Id: Ia773de475431eb85fb0dbe724d524e8dd59b1b12
Reviewed-on: http://gerrit.cloudera.org:8080/5191
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>