encodings: change IsBlockFull() to not take a limit parameter
Everywhere we called this, we used the same value (the configured cfile
block size). Given that, it's simpler to not take the parameter. This
makes it easier for an encoder to determine whether or not it is full in
some other context, since it can safely assume that it knows the block
size up front.

Change-Id: I4598c1555732649eeebf43231082dd641ecfc576
Reviewed-on: http://gerrit.cloudera.org:8080/5200
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>