cfile-test: use a faster data generator for 100M-string test
The current data generator using StringPrintf is not very fast, and
since we don't care about the contents of the file for this test, it's
just as easy to use FastHex64ToBuffer for better performance.

This resulted in a ~2x speedup in the test. This means that the test
will now better represent the underlying performance of the encoder and
cfile writing code.

Change-Id: I9578681aa2e064dbecd79fdce7b8a083ef03b3c4
Reviewed-on: http://gerrit.cloudera.org:8080/5201
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>