dist-test: set TEST_TMPDIR inside test directory
Recently I hit a case where I had a test which was crashing with an ASAN
warning halfway through, after writing a bunch of data into the test
tmpdir. Because we weren't setting TEST_TMPDIR in the dist-test
environment, the leftover data ended up orphaned in /tmp/kudutest-1000,
which wasn't cleaned up after the dist-test task ended.

This just sets the tmpdir to be inside the root working directory of the
test, so that we are guaranteed to clean it up after the test finishes.

To verify, I submitted a job and looked at the resulting test logs, and
saw that it was now writing within the test directory instead of within
/tmp/

Change-Id: I0497591860f7e8b1ea18c1ee9b05becb282f6ec8
Reviewed-on: http://gerrit.cloudera.org:8080/5203
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins