Note to install "which" command in the installation guide in case of building on RHEL/CentOS.
I tried to build on Docker container with official CentOS7 image and 
when I tried to build thirdparty libraries, I caught an error message as follows.

+ /root/work/kudu/thirdparty/preflight.py
Running pre-flight checks
-------------------------
Using C compiler: c++
Using C++ compiler: c++

  (Set $CC and $CXX to change compiler)
-------------------------
Checking for autoconf
Traceback (most recent call last):
  File "/root/work/kudu/thirdparty/preflight.py", line 149, in <module>
    sys.exit(main())
  File "/root/work/kudu/thirdparty/preflight.py", line 141, in main
    check_tools()
  File "/root/work/kudu/thirdparty/preflight.py", line 97, in check_tools
    if subprocess.call(["which", tool], stdout=DEV_NULL, stderr=DEV_NULL) != 0:
  File "/usr/lib64/python2.7/subprocess.py", line 524, in call
    return Popen(*popenargs, **kwargs).wait()
  File "/usr/lib64/python2.7/subprocess.py", line 711, in __init__
    errread, errwrite)
  File "/usr/lib64/python2.7/subprocess.py", line 1327, in _execute_child
    raise child_exception
OSError: [Errno 2] No such file or directory

This message means that "which" command is absent.
Of course, I understand it's very rare case that "which" is absent but I think 
it's better to note just in case to install "which" command in the installation guide.

Change-Id: If02eae6f22a51965120a037cc9a68ddefa418b6c
Reviewed-on: http://gerrit.cloudera.org:8080/5216
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>