Remove/downgrade a few very verbose useless log statements
When running tests that use a tablet server the log is
full of thread/reactor/maintenance manager statetments
that are useless. This is particularly annoying at VLOG(5)
level where we expect to see printouts of compaction
inputs/outputs. These get lost in a sea of useless statements.

This outright removes the "timer tick" reactor statement,
downgrades another reactor statement to VLOG(10) and makes the
maintenance_manager much quieter when it's disabled.

Change-Id: I7863f067f49cc7432e8ca3b532a208d3c4005300
Reviewed-on: http://gerrit.cloudera.org:8080/5226
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins