Change TabletHistoryGCITest back to manual flushes
Mike found a potential problem with client flushes when there are
concurrent manual and automatic flushes going on (KUDU-1761). Until
we address that this changes this test back to MANUAL_FLUSH with
a single flush at the end of each batch.

This change makes the test non-flaky (ran dist-test on a follow
up patch).

Change-Id: Ibf89a8b113f347a274aa59e59e0d8b7ebcb6f2f0
Reviewed-on: http://gerrit.cloudera.org:8080/5224
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>