Reduce the number of batches in FuzzTest::TestFuzzHugeBatches
While running fuzz-itest in dist-test with slow mode, asan, it's
often flaky due to timeouts. This test is the main culprit
as it often takes 13 minutes to complete.

This reduces the number of operations in this test and
'update_multiplier' to half for all build types but TSAN (we had already
reduced ops in this case) making it non-flaky.

Change-Id: Ib73220ed342d2417ad65bf1ae705499cab7a9b10
Reviewed-on: http://gerrit.cloudera.org:8080/5223
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins