KUDU-1773: remove overly strict DCHECKs
A tablet's replicas may change over the course of a write. For example,
between two attempts at the same write, a replica could be evicted from
the replication group and replaced by another.

A pair of DCHECKs in the metacache didn't allow that, so they've been
removed. Furthermore, their respective RemoteTablet functions now don't
return anything, to reflect the fact that success or failure in finding the
replica in question should be immaterial to all callers.

Change-Id: I01c1dde99cce1f43e6a9864c1ff6f7aaad448a77
Reviewed-on: http://gerrit.cloudera.org:8080/5292
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>