Ignore tablet metadata files with old tmp infix
Commit 8ff4d1211d2a9f20203774bd263112ec0d7333e1 changed the tmp infix
from '.tmp' to '.kudutmp'. However, if we crash a server on an old
version and then restart with the new version, the server won't know to
ignore the tmp files from the old version in the tablet metadata
directory.

This will result in failures to start up as it tries to read these tmp
files as if they were completed tablet metadata files.

Change-Id: Ie23d6d989472ec38b36d6a2f6d3f00e32df22869
Reviewed-on: http://gerrit.cloudera.org:8080/5298
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>