[java client] Allow to pass extra flags to tservers/masters
This path adds a couple of methods to MiniKuduClusterBuilder so that
a test can pass extra flags to the daemons. Another change is that
the builder is now a protected static final field of BaseKuduTest
so that subclasses can set flags before starting the cluster.

This is required for TestHybridTime to pass for a follow up patch.

Change-Id: If9ecf0d22c0b49e2a07d5e3cf21c762e0790361a
Reviewed-on: http://gerrit.cloudera.org:8080/5299
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>