KUDU-798 (part 4) Add a TimeManager to manage safe time advancement
This patch adds an entity to manage safe time (the timestamp before
which all transactions are committed/aborted or in-flight) across
a consensus configuration. This allows scans to make sure/wait that
the scan timestamp is safe, thus making sure the scan is repeatable.

This adds a small unit test. Follow up patches will use this
in integration tests.

Change-Id: I0bb2b1d2590ed7ead6f1980f9572be10444bb81b
Reviewed-on: http://gerrit.cloudera.org:8080/5300
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>