KUDU-1778. Fix LMP mismatch behavior after a replica restarts
This fixes an issue seen in a stress test after a cluster restart. Both
replicas had an LMP mismatch with the leader, and the tablet was unable
to make progress.

The issue turned out to be that the followers were returning 0 as their
committed index, and the leader then tried to fall back to index 1. That
index had already been GCed, and thus the leader was unable to send any
operations to the followers.

I tested this patch in the same stress test environment and the issue
didn't reproduce. This also includes a test which failed without the
fix. I looped the new test 500 times and it passed.

Change-Id: I8f1332d605f7f846a01923b3ab92f12d73462bba
Reviewed-on: http://gerrit.cloudera.org:8080/5309
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>